Extracted some common utils to separate file and added funcionality to fix overflow consumers accounts
